restructured login code, consolidated routes, added login link on nav bar
