+++ views/layouts/main.hbs
@@ -23,11 +23,13 @@
     <div class="logo">{{> logo}}</div>
     <div class="nav">{{> nav-bar}}</div>
     {{{body}}}
+    {{> modal}}
     <script
       src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
       integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
       crossorigin="anonymous"
     ></script>
     <script src="/js/script.js"></script>
+    <script src="/js/login-script.js"></script>
   </body>
 </html>
